protect against mistakes in DEV
We currently have the power to edit any metadata of any video in our account in DEV. This is scary!

This PR adds some protection so that we don't, for example, change the title of a video of a public video to something like "test". Instead, we can only edit the metadata of video in a channel in config.youtube.allowedChannels.