endpoint for Commercial diagnosis
As part of the Commercial training for selling YouTube, one of the troubleshooting topics is: "Why isn't this video showing in DfP?".

One way to answer that is to login to the YouTube CMS and find the video. However this requires adding users to the CMS. This endpoint will call the YouTube Partner API and display the advertising options for a video and error if the video isn't claimed or isn't owned by the content owner.

The endpoint is `/api/youtube/video-info/:id` and is Panda Authed.

As part of this, I've refactored YouTubeClaims as a trait, following the YouTubeVideos pattern. I've also simplified the `YouTubeChannel` case class, removing `logo` as its never used.