fix usage bug on create form
Usage aren't getting updated when you create a new atom. This is ok when there is nothing in the store (blankUsageData), however if you've looked at an atom, then go to create a new atom, the usages do not change, that is, the usages of the newly created atom are the same as the atom you were previously looking at.

This fixes this by refreshing usages when creating a new atom.