fix tooltip opacity issue
fix tooltip opacity issue by putting it further up the DOM. The opacity style effects all children, before, the tooltip was a child of an element whose opacity was set to 0.4 when it had a disabled class. Now, the tooltip is higher up the DOM so doesn't inherit any opacity.