reduce character count
Youtube only allow 500 chars in the tags field. For some reason, we're exceeding that even though we check it client side. I think this is because of the content bundle tag that we add server side.

Reducing the client side limit means we reserve space for the server side addition.