refactor duration and the PublishAtomCommand
Calculating the duration requires a call to YouTube. Whilst this isn't a problem in general, making this call during publishing feels unnecessary as we can do it during asset activation. So, refactor ActiveAssetCommand too, simplifying the routes of entry; you only need an atom id and a version number to activate an asset.

The side-effect of this, is one less call to YouTube, reducing our API quota usage and also a slightly faster publishing route as there's one less call to a third party API.