change atom channel on add asset
There have been a couple of times where Adam has wanted to change the channel of an Atom because of incorrect selection but was unable to. In these scenarios, we either curl the update endpint, editing the channel, or create a new atom. Both aren't great.

This PR changes the behaviour of the add asset endpoint allowing you to paste in the URL of any video and if the channel is recognised, the Atom's channel gets updated to match.