send an email when a new YouTube video is uploaded with WC tags
We've got an ad campaign for the WC which will get served when a video has the tag `world cup 2018` or `2018 world cup`.

There have been a couple of instances where the YouTube video didn't make it to DfP for targeting. We're not sure why this is as the video appears to be tagged correctly on youtube.com.

Tom suggested sending him an email when a new video is uploaded to YouTube with these tags so he can then check DfP immediately rather than finding DfP is missing items when its too late.

Note: the email gets sent when an atom is published with a new YouTube asset, or the tags have been added to the existing asset.