render drop down as a dropdown
Currently, there is some confusion around this button (some reports of not daring to click on it!). I think this is due to the fact it looks like a button rather than a drop down.

This is a quick attempt at making it more obvious. Design inspired by Composer's similar component.