Reduce the number of calls to the YouTube Partner API when making a claim
YouTube's Partner API is rate limited on the number of calls that can be made in a day and also in a second. We've been seeing some API calls being rejected because `userRateLimitExceeded`.

This changes the way we find a previous claim on a video by calling `setOnBehalfOfContentOwner`, `setIncludeThirdPartyClaims` and `setPartnerUploaded` to only return first party claims. This should return 0 or 1 items. This significantly reduces the number of API calls we're doing - previously, we were getting *all* claims and then filtering in scala. This should result is us hitting `userRateLimitExceeded` less often.