remove created date from pluto projects
We have had multiple reports of commissions and projects not appearing in MAM. Looking at the logs, they are littered with:

```
Caused by: java.lang.IllegalArgumentException: Invalid format: "2018-09-11T10:54:14.582880" is too short
	at org.joda.time.format.DateTimeFormatter.parseDateTime(DateTimeFormatter.java:945)
```

I've not been able to fix the issue upstream in Pluto. Dates are hard and we're not using this field so lets remove it and get things working again.