remove the /videos/pluto-list endpoint
On PROD, the API that powers this is returning 891 items at a weight of 3.6MB... I suspect this page isn't being used if there are so many outstanding items.

Removing this page keeps the repo clean and also means we can reduce load on dynamo as the API was doing a full table scan.