fail healthcheck if we're unable to read stage tag
We use the stage tag to determine what config and ultimiately what resources to use. If the stage isn't available, we're defaulting to `DEV`, which means PROD tries to access DEV resources.

That's not good!

Now we're failing to report healthy if we're running in AWS and we're reporting stage as DEV. This will have the effect of a deploy succeeding only when we're able to correctly read the stage tag.