set page-size param when asking for usages
The default value of `page-size` is 10, however the atom `7181c82d-f06e-481e-9368-c70de70396ce` in PROD has 15 usages, this means the app isn't showing the correct data.

This is a bit of a cheat to get around the issue; ideally we'd iterate over the pages. However that is a bigger change than this and would create more complex code for not much more benefit in the immediate term.