fix: re-id the blockAds DOM element
Although we previously rejected this change in #574,
we've just had an issue with a shopping browser extension (https://chrome.google.com/webstore/detail/honey/bmnlcjabgnpnenekpadlanbbkooimhnj?hl=en).

Some extensions are removing the checkbox for the block ads field from the DOM because they think it is an advert.

This change changes the DOM element's ID to a less vulnerable one, hopefully preventing ad-blockers and other browser extensions from finding it,
creating an application that is usable in more situations.