Improve layout for video images and provide descriptions
With a third video field (the youtube override image) there is a need
to make a better use of the space in the interface.

It also seems sensible to explain what the three images do, and
especially their precedence - i.e. that the thumbnail and youtube
images are optional overrides.