Fix unexpected behaviour
getByPath takes two arguments, the second is a boolean `retry`. The map method in JS takes a callback function that accepts three arguments, `element`, `index` and `array`. Typically `index` and `array` are ignored by the callback functions, but in this case `index` will be interpreted as `retry`. So for the first element in the array `retry` will be zero (`false`) but for subsequent items it will be non-zero (`frue`). Which is probably unexpected.