Surface revision errors rather than masking them
Sets revision ID based on what the client has sent, not what we've just fetched from dynamodb. This way, if the user has edited an old version of the atom, atom-publisher-lib will refuse to update the db. Atom-publisher-lib will only publish if the new revision is higher than the current revision.