Pre-Play 2.9: Upgrade pan-domain-authentication (Panda) library
This is in preparation for the Play 2.9 upgrade in
https://github.com/guardian/media-atom-maker/pull/1140 - the current
Panda artifacts used actually _do_ have Scala 2.13 versions at the
current artifact version number:

- https://repo1.maven.org/maven2/com/gu/pan-domain-auth-play_2-8_2.13/1.2.0/
- https://repo1.maven.org/maven2/com/gu/panda-hmac-play_2-8_2.13/2.1.0/

...but they _don't_ have Play 2.9 versions, not at this artifact number -
so for the later upgrade to Play 2.9, we will need to update the Panda
artifact version number, and it makes sense to separate this change out
to make the later change smaller.

Note that the panda-hmac-play artifacts have been moved out
of the archived https://github.com/guardian/panda-hmac repo,
with the last version there being v2.2.0 released in June 2023.
They are now in the main https://github.com/guardian/pan-domain-authentication
repo, and releases are now done with the other panda artifacts -
so the version numbers of all the panda artifacts, including
panda-hmac, are now aligned.