Add unit test for YouTube client initialisation
In the past, we've had runtime failures when we've updated dependencies
(like Panda in https://github.com/guardian/media-atom-maker/pull/1123)
that update the version of google-api-client library used, without
also updating the YouTube client SDK to match (note, for the reasons below,
it can be hard to get updated compatible versions of the YouTube client).
The YouTube client specifically checks whether it is using an expected
version of the google-api-client library, and will throw an exception like
this if its requirements aren't met:

```
java.lang.IllegalStateException: You are currently running with version 2.1.2 of google-api-client. You need at least version 1.15 of google-api-client to run version 1.22.0 of the YouTube Data API library.
```

This change adds a unit test that just exercises the constructor of both
the `YouTube` & `YouTubePartner` clients, and which will fail in the same
way the client did at runtime - but, better for our purposes, it will fail
the unit test, not the running app server!

An example of the test failing is here, against PR #1145:

https://github.com/guardian/media-atom-maker/actions/runs/7828479116/job/21358445038#step:8:3721

# Getting updated compatible versions of the YouTube client

Media Atom Maker uses the semi-private YouTube content partner API. From an SDK
point of view, this involves two particular clients, which we started using as
of https://github.com/guardian/media-atom-maker/pull/399 in April 2017:

* `com.google.api.services.youtube.YouTube` - a _modified_ version of the standard client,
  with additional fields (like `onBehalfOfContentOwner`) that are visible in Google's
  APIs Discovery Service, but _not_ typically included in the standard
  `google-api-services-youtube` client SDK published to Maven Central:
  https://central.sonatype.com/artifact/com.google.apis/google-api-services-youtube/versions
* `com.google.api.services.youtubePartner.YouTubePartner` - an _additional_ client, that
  hits a different API endpoint and has different actions entirely from the standard
  `YouTube` client (claims, etc).

In the past, Google _would_ share special jars for both clients, which were hosted on
their docs pages at:

https://developers.google.com/youtube/partner/client_libraries

However, that page now links off to just the standard versions of the
Google API Client Library for Java - with the unmodified `YouTube` client,
and no `YouTubePartner` at all.

Consequently, we have https://github.com/guardian/media-atom-maker/pull/1129,
which scripts the production of the custom Content Partner API clients.

* https://github.com/guardian/media-atom-maker/pull/1145#issuecomment-1931940022
* https://github.com/guardian/media-atom-maker/pull/1124
* https://github.com/guardian/media-atom-maker/pull/1129