Wait for server before updating active asset
This commit updates the client to wait on a successful response from the
server (when the user selects a new asset to be active) before marking
that asset as active in the client state.

Before now, the client-side code would immediately update which asset
was marked as active before sending the request to the server: see
[revertAsset](https://github.com/guardian/media-atom-maker/blob/47c84b2e321fd1c778d530e5fbf884896f79d910/public/video-ui/src/actions/VideoActions/revertAsset.js#L32-L35).

This meant that if the request failed, the client’s state had updated to
show the intended asset as active, while the server’s state had not. We
suspect this caused an error where the wrong video was published.

This code was added in
https://github.com/guardian/media-atom-maker/pull/92, where it’s not
currently clear whether the choice to use ASSET_REVERT_REQUEST was
deliberate or an accident: I’m guessing the latter only because I can’t
think of a rationale for doing it this way.

Are there any ways we could have caught this error, perhaps with some
testing? Are there ways we could find other occurrences of similar
errors? Could we update the app to discard writes from clients with
information that is out of sync with the server?