Fix getStatus
When copying the old definition I accidentally destructured where I
shouldn’t have, so the promise was always rejecting with a destructuring
error.

However, this does highlight a problem here: I got no type error telling
me this wouldn’t work, because both the WorkflowApi and the caller of
getStatus are untyped.