Migrate localUpdateWorkflowData and swap reducers
Until now I’ve left both reducers in, but that meant that the state from
the new one wasn’t being read by the components, which were still
looking at the old state. That state was no longer being updated because
I switched to dispatching the new actions, which meant editing the
workflow form was broken.