Tighten new reducer types
Use empty object as initial state for status

I changed this to null, but that actually broke the app: it’s written to
expect an empty object when data hasn’t been fetched from workflow yet.
I think we should probably change it to handle null correctly, as it’s
more clear, but for now I don’t want to do that.