Restore _getResponseAsJson
I removed this originally because I’d thought that apiRequest would
automatically decode response bodies as json. However, that actually
only happens if the response includes a content type header whose type
is application/json, which the workflow-frontend endpoints don’t!

So this commit restores the json parsing, which is required.

We should update the workflow-frontend endpoints to specify the
content-type they’re returning to be json, though.