Add types to WorkflowApi
This commit adds types for the workflow api, based on inspecting the
scala code in workflow-frontend and workflow, as well as the responses
currently returned to media-atom-maker-CODE by workflow-CODE.

The types are somewhat complicated, which is mildly frustrating: it
seems unnecessary for workflow-frontend to flatten the stubs like it
does, for example, and it means one has to read the relevant Encoder
there to understand the structure of the data returned by the API. It
would be nicer to have that type officially declared, and possibly just
leave it as the Stub type – oh well.

I wonder would it be nice to define the API datatypes in a way that the
typescript types (and probably zod schemas) could be automatically
derived from them, like we do in some places with thrift? It would save
needing to redefine these types here and in composer, but it would
introduce a new dependency and require more synchronisation, as we’ve
seen with the thrift dependencies.

Originally this commit also removed _getResponseAsJson, because I’d
thought that apiRequest would automatically decode response bodies as
json. However, that actually only happens if the response includes a
content type header whose type is application/json, which the
workflow-frontend endpoints don’t!

I think we should possibly update the workflow-frontend endpoints to
specify the content-type they’re returning to be json, at which point we
could remove _getResponseAsJson again.