WIP: fix embargo date handling
This code does not all handle the embargo dates correctly! From testing
in CODE, the embargo value in the redux store is an ISO date string:
this matches with the definition of VideoUtils.getEmbargoAsDate(), which
clearly makes that assumption.

However, some parts of the code currently compare the ISO date string to
impossiblyDistantDate, which is a number of milliseconds since the epoch
– this is unlikely to give the desired result.