Account for null and undefined in types
Copilot identified on the PR that some things might be null: I hadn’t
seen this already because [strict
mode](https://www.typescriptlang.org/docs/handbook/2/basic-types.html#strictness)
is not currently enabled in our typescript config: enabling it and
making these changes gets rid of the errors in this file.

The resulting types for some fields of ExternalData are a little
unpleasant: we’ve got `undefined` from the optionality of the field, but
we also have `null`. However, `null` is definitely already used in the
code, and the field can be absent, so while it’s a little messy I
believe it’s correct.