Restore old behaviour of _getResponseAsJson
Also rename it to something clearer.

I originally figured we shouldn’t lie to the typechecker here, and
should instead throw an error on a non-string type returned. However, it
was pointed out to me that writing it this way is a handy way to prevent
it from breaking if the workflow API starts to return items with an
application/json content-type. Since I do plan to make that change, this
seems sensible: once we’ve done that we can remove this function again.