Try elk renderer for architecture diagram
The mermaid docs [describe the option of using a second
renderer](https://mermaid.js.org/syntax/flowchart.html#renderer), which
they say is more suitable for large, complex diagrams. This is
definitely a large, complex diagram, so let’s give it a shot!

Is this the right way to configure it though? Not sure: gonna push this
commit to github to test.